finish minishell

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: melisha <melisha@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 11:28:05 by melisha           #+#    #+#             */
-/*   Updated: 2021/03/17 11:28:23 by melisha          ###   ########.fr       */
+/*   Updated: 2021/03/18 21:25:23 by melisha          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libminishell.h"
 
-size_t		ft_strlen(const char *str)
+int		ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
